refactor: [#56] extract mods in upgrader
