+++ src/upgrades/from_v1_0_0_to_v2_0_0/mod.rs
@@ -1,2 +1,3 @@
-pub mod upgrader;
 pub mod databases;
+pub mod transferrers;
+pub mod upgrader;
