+++ src/upgrades/from_v1_0_0_to_v2_0_0/transferrers/mod.rs
@@ -1,0 +1,4 @@
+pub mod category_transferrer;
+pub mod torrent_transferrer;
+pub mod tracker_key_transferrer;
+pub mod user_transferrer;
